Refactoring + new embedded chat:
* moving conversejs plugin in separate files
* disconnecting embedded chat on navigation-end
