+++ conversejs/index.html
@@ -14,10 +14,11 @@
 <noscript>You need to enable JavaScript to run the Converse.js chat app.</noscript>
 <div id="conversejs-bg" class="theme-peertube"></div>
 <script type="text/javascript">
+  initConversePlugins(false);
   initConverse(
     {INIT_CONVERSE_PARAMS},
     'chat-only'
-  )
+  );
 </script>
 </body>
 </html>
